move query and fix loading state
